requirements.txt
# local package
-e .

# external requirements
click
Sphinx
coverage
awscli
flake8
python-dotenv>=0.5.1
SQLAlchemy==1.3.16
pandas==1.0.3 
